fix: nasty sentence over-writing bug
In the end the bug was due to a collision in the sentences_by_end map.
What happened is when editing at certain times a sentence was overwriting
an old one because luck would have it they had the same end loc (which is used
as a key in the map in question). This would then have two effects:
- Remove it in the sentences to end map
- Duplicating it in the sentences by id map as in subsequent passes it was no
longer treated